requirements.txt
arabic-reshaper==1.0.1
Django==1.10.4
MySQL-python==1.2.5
Pillow==3.4.2
python-bidi==0.4.0
reportlab==3.3.0
six==1.10.0
